ascotraceR: Simulate the Spread of Ascochyta Blight in Chickpea

A spatiotemporal model that simulates the spread of Ascochyta blight in chickpea fields based on location-specific weather conditions. This model is adapted from a model developed by Diggle et al. (2002) <doi:10.1094/PHYTO.2002.92.10.1110> for simulating the spread of anthracnose in a lupin field.

See also

Author

Maintainer: Paul Melloy p.melloy@uq.edu.au (ORCID) [translator]

Authors:

Ihsanul Khaliq ihsan.khaliq@usq.edu.au (ORCID)

Adam H. Sparks adamhsparks@gmail.com (ORCID) [conceptor]

Other contributors:

Grains Research and Development Corporation (GRDC) Project USQ1903-003RTX [funder, copyright holder]

University of Southern Queensland [copyright holder]

Western Australia Agriculture Authority (WAAA) (Supported the development of ascotraceR through Adam H. Sparks' time.) [copyright holder]